+++ config-example-ttn.js
@@ -87,6 +87,7 @@ module.exports = {
     //     }
     // },
 
+    // Optional watchdogs, only reported to Slack and/or Telegram.
     watchdogs: [
         {
             // No include/exclude patterns, hence the pure heartbeat watchdog that is satisfied for any UART byte
@@ -101,6 +102,7 @@ module.exports = {
             // matched for a FULL line from the UART. To independently monitor multiple types of messages one should
             // configure multiple watchdogs, instead of listing multiple patterns within a single watchdog.
             name: 'LoRaWAN uplink',
+            // MQTT: Sending UPLINK OK
             include: [/sending uplink ok/i],
             exclude: [],
             timeout: 15 * MINUTES,
@@ -109,6 +111,7 @@ module.exports = {
         {
             // The TTN Gateway sends a status message every 30 seconds, but booting might take a bit longer
             name: 'gateway status',
+            // MQTT: Sending status succeeded
             include: [/sending status succeeded/i],
             exclude: [],
             timeout: 5 * MINUTES,
@@ -117,11 +120,122 @@ module.exports = {
         {
             // The TTN Gateway reboots for a firmware upgrade once per day
             name: 'daily firmware upgrade',
-            include: [/rebooting gateway for firmware update check/i],
+            // MAIN: Rebooting gateway for firmware update check
+            include: [/firmware update/i],
             exclude: [],
             timeout: 25 * HOURS,
             repeat: 12 * HOURS
         }
+    ],
+
+    // Optional reporters, only sending notifications to Slack and/or Telegram.
+    reporters: [
+        {
+            // One can define multiple schedules, where each will keep track of its own counts for its interval.
+            schedules: [
+                // Enable for quick testing:
+                // {
+                //     name: '5 minutes report',
+                //     interval: 5 * MINUTES,
+                //     level: 'warn'
+                // },
+                {
+                    name: 'Hourly gateway report',
+                    // The interval for which to collect and report the statistics. The very first report might use a
+                    // smaller interval, as the monitor runs the reporting at recognizable times. Like if the interval
+                    // is an exact number of days, the first report is sent to Slack and/or Telegram around midnight
+                    // (LOCAL TIME), no matter when the monitor was started. Likewise this detects multiples of hours,
+                    // 30 minutes, 15 minutes, 10 minutes, 5 minutes, 1 minute, 30 seconds, 15 seconds, 10 seconds,
+                    // 5 seconds and a single second.
+                    interval: 60 * MINUTES,
+                    // The log level at which to report the statistics, matched against the minimum level for Slack and
+                    // Telegram. Here: Slack only.
+                    level: 'warn'
+                },
+                {
+                    name: 'Daily gateway report',
+                    interval: 24 * HOURS,
+                    // Both Slack and Telegram
+                    level: 'error'
+                }
+            ],
+            reports: [
+                {
+                    name: 'dropped LoRa packets',
+                    // LORA: Packet dropped! Bad CRC
+                    include: [/LORA: Packet dropped/i],
+                    exclude: []
+                },
+                {
+                    name: 'accepted LoRa packets',
+                    include: [/LORA: Accepted packet/i],
+                    exclude: []
+                },
+                {
+                    name: 'forwarded LoRa uplinks',
+                    include: [/MQTT: Sending UPLINK OK/i],
+                    exclude: []
+                },
+                {
+                    name: 'received LoRa downlinks',
+                    include: [/MQTT: Received DOWNLINK/i],
+                    exclude: []
+                },
+                {
+                    name: 'status packets',
+                    include: [/MQTT: Sending status packet/i],
+                    // MQTT: Sending status succeeded: 11
+                    exclude: [/MQTT: Sending status succeeded/]
+                },
+                {
+                    name: 'MQTT errors',
+                    // MQTT: Connection failed
+                    // MQTT: Sending status failed
+                    // MQTT: Sending UPLINK failed: -1
+                    // ...all followed by:
+                    // MAIN: MQTT error
+                    include: [/MQTT.* error/i],
+                    // MQTT: Report reboot error: 0110
+                    exclude: [/report .* error/i]
+                },
+                {
+                    name: 'network errors',
+                    // INET: Connected to a network, waiting for DHCP lease, checking validity with ping
+                    // INET: Trying to connect to WiFi router again (after 120 seconds)
+                    // INET: No Ethernet and WiFi link (after 10 seconds)
+                    include: [/INET: No Ethernet and WiFi link/i],
+                    // INET: Error sending probe on Eth
+                    // INET: Error sending probe on WiFi
+                    exclude: [/INET: Error sending probe/]
+                },
+                {
+                    name: 'reboots',
+                    // *** Application reboot
+                    // Reboot reason: 0x10
+                    include: [/Reboot reason/i],
+                    // MAIN: Rebooting gateway for firmware update check
+                    exclude: [/upgrade check/i]
+                }
+            ]
+        },
+        {
+            schedules: [
+                {
+                    name: 'Hourly UART report',
+                    interval: 60 * MINUTES,
+                    // Slack only
+                    level: 'warn'
+                }
+            ],
+            reports: [
+                {
+                    name: 'UART messages',
+                    include: [/.*/i],
+                    exclude: []
+                }
+            ]
+        }
+
     ]
 
 };